# backend/routers/jobs.py
from fastapi import APIRouter, UploadFile, File, Form, HTTPException, BackgroundTasks
from pathlib import Path
import uuid
import os
import json
from datetime import datetime
from schemas.schemas import (
    ProcessConfig, VizConfig, JobCreateResponse, EvaluateJobRequest, TestPair,
)
from services.video_processor import process_video, request_cancel
from services.evaluation.evaluator import run_evaluation
from services.evaluation.gt_loader import list_available_pairs
from db.mongodb import get_db

router = APIRouter(prefix="/jobs", tags=["Jobs"])

@router.post("", response_model=JobCreateResponse)
async def create_job(
    background_tasks: BackgroundTasks,
    video: UploadFile = File(...),
    config: str = Form(...),
    viz_config: str = Form(...)
):
    try:
        cfg = json.loads(config)
        v_cfg = json.loads(viz_config)
    except json.JSONDecodeError:
        raise HTTPException(status_code=400, detail="Invalid JSON in config or viz_config")

    from config import BASE_DIR
    import hashlib

    job_id = str(uuid.uuid4())
    upload_dir = str(BASE_DIR / "data" / "videos" / "uploads")
    os.makedirs(upload_dir, exist_ok=True)
    
    content = await video.read()
    file_hash = hashlib.md5(content).hexdigest()
    
    file_path = f"{upload_dir}/{file_hash}_{video.filename}"
    if not os.path.exists(file_path):
        with open(file_path, "wb") as f:
            f.write(content)

    db = get_db()
    
    job_doc = {
        "job_id": job_id,
        "filename": video.filename,
        "created_at": datetime.utcnow(),
        "status": "pending",
        "progress": 0.0,
        "config": cfg,
        "viz_config": v_cfg,
        "live_stats": {},
        "result": None,
        "error": None,
        "input_path": file_path,
    }
    await db["jobs"].insert_one(job_doc)

    background_tasks.add_task(process_video, job_id, file_path, cfg, v_cfg)
    
    return {"job_id": job_id, "status": "pending"}

@router.get("/{job_id}")
async def get_job(job_id: str):
    db = get_db()
    job = await db["jobs"].find_one({"job_id": job_id}, {"_id": 0})
    if not job:
        raise HTTPException(status_code=404, detail="Job not found")
    return job

from pydantic import BaseModel
class TestJobRequest(BaseModel):
    filename: str
    config: ProcessConfig
    viz_config: VizConfig

@router.post("/test", response_model=JobCreateResponse)
async def create_test_job(
    request: TestJobRequest,
    background_tasks: BackgroundTasks
):
    from services.evaluation.gt_loader import gt_root
    
    # filename format: "{basename}/video.mp4"
    source_path = gt_root() / request.filename
    if not source_path.exists():
        raise HTTPException(status_code=404, detail=f"Test video not found: {request.filename}")
        
    job_id = str(uuid.uuid4())
    file_path = str(source_path)
    
    db = get_db()
    
    job_doc = {
        "job_id": job_id,
        "filename": request.filename,
        "created_at": datetime.utcnow(),
        "status": "pending",
        "progress": 0.0,
        "config": request.config.model_dump(),
        "viz_config": request.viz_config.model_dump(),
        "live_stats": {},
        "result": None,
        "error": None,
        "input_path": file_path,
    }
    await db["jobs"].insert_one(job_doc)

    background_tasks.add_task(process_video, job_id, file_path, request.config.model_dump(), request.viz_config.model_dump())
    
    return {"job_id": job_id, "status": "pending"}

@router.get("/test/videos")
async def list_test_videos():
    """Список тестових відео з GT-директорії (для Upload → демо-запис)."""
    from services.evaluation.gt_loader import gt_root
    root = gt_root()
    if not root.exists():
        return []
    videos = []
    for subdir in sorted(root.iterdir()):
        if not subdir.is_dir():
            continue
        video = subdir / "video.mp4"
        if video.exists():
            videos.append(f"{subdir.name}/video.mp4")
    return videos


@router.get("/evaluate/test-pairs", response_model=list[TestPair])
async def list_evaluation_pairs():
    """Доступні відео+GT пари з GT_DATASET_PATH."""
    return list_available_pairs()


@router.post("/evaluate", response_model=JobCreateResponse)
async def create_eval_job(
    request: EvaluateJobRequest,
    background_tasks: BackgroundTasks,
):
    from services.evaluation.gt_loader import gt_paths

    paths = gt_paths(request.gt_basename)
    video_path = paths["video"]
    if not video_path.exists():
        raise HTTPException(
            status_code=404,
            detail=f"GT-відео для '{request.gt_basename}' не знайдено: {video_path}",
        )

    job_id = str(uuid.uuid4())
    file_path = str(video_path)

    db = get_db()
    job_doc = {
        "job_id": job_id,
        "filename": request.gt_basename,
        "created_at": datetime.utcnow(),
        "status": "pending",
        "progress": 0.0,
        "config": request.config.model_dump(),
        "viz_config": request.viz_config.model_dump(),
        "live_stats": {},
        "result": None,
        "error": None,
        "input_path": file_path,
        "evaluation": None,
        "gt_basename": request.gt_basename,
    }
    await db["jobs"].insert_one(job_doc)

    background_tasks.add_task(
        run_evaluation,
        job_id, file_path, request.gt_basename,
        request.config.model_dump(), request.viz_config.model_dump(),
    )

    return {"job_id": job_id, "status": "pending"}

@router.delete("/{job_id}")
async def delete_job(job_id: str):
    db = get_db()
    job = await db["jobs"].find_one({"job_id": job_id})
    if not job:
        raise HTTPException(status_code=404, detail="Job not found")

    # Signal running processor to stop (no-op if job is not active)
    request_cancel(job_id)

    await db["jobs"].delete_one({"job_id": job_id})

    from config import settings
    output_dir = Path(settings.OUTPUT_DIR)

    for fname in [f"{job_id}_raw.mp4", f"{job_id}.mp4"]:
        (output_dir / fname).unlink(missing_ok=True)

    input_path = job.get("input_path")
    if input_path and "uploads" in input_path:
        # Delete the file only if no other job is using it
        other_jobs = await db["jobs"].count_documents({"input_path": input_path, "job_id": {"$ne": job_id}})
        if other_jobs == 0:
            Path(input_path).unlink(missing_ok=True)

    return {"status": "deleted"}

@router.get("/{job_id}/live")
async def get_job_live_stats(job_id: str):
    db = get_db()
    job = await db["jobs"].find_one({"job_id": job_id}, {"live_stats": 1, "status": 1, "progress": 1, "_id": 0})
    if not job:
        raise HTTPException(status_code=404, detail="Job not found")
    return job

@router.get("")
async def list_jobs():
    db = get_db()
    jobs_cursor = db["jobs"].find({}, {"_id": 0, "events": 0}).sort("created_at", -1).limit(20)
    jobs = await jobs_cursor.to_list(length=20)
    return jobs